+++ lib/manager/npm/index.ts
@@ -2,9 +2,12 @@ import { LANGUAGE_JAVASCRIPT } from '../../constants/languages';
 import * as npmVersioning from '../../versioning/npm';
 
 export { extractAllPackageFiles } from './extract';
-export { bumpPackageVersion, updateDependency } from './update';
+export {
+  bumpPackageVersion,
+  updateDependency,
+  updateLockedDependency,
+} from './update';
 export { getRangeStrategy } from './range';
-export { updateLockedDependency } from './update-locked';
 
 export const language = LANGUAGE_JAVASCRIPT;
 export const supportsLockFileMaintenance = true;
