+++ lib/manager/npm/update/index.ts
@@ -1,2 +1,3 @@
-export { bumpPackageVersion } from './package-version';
 export { updateDependency } from './dependency';
+export { updateLockedDependency } from './locked-dependency';
+export { bumpPackageVersion } from './package-version';
