+++ lib/manager/npm/update/locked-dependency/__fixtures__/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "express": "4.0.0"
+  }
+}
